+++ src/bin/drone-ld.rs
@@ -4,13 +4,17 @@ use anyhow::Result;
 use drone::{
     color::Color,
     templates::Registry,
-    utils::{block_with_signals, register_signals, run_command, run_wrapper, search_rust_tool},
+    utils::{
+        block_with_signals, crate_root, register_signals, run_command, run_wrapper,
+        search_rust_tool,
+    },
 };
 use drone_config::Config;
 use std::{
     collections::HashMap,
     env,
     ffi::{OsStr, OsString},
+    fs::{create_dir_all, File},
     path::Path,
     process::Command,
 };
@@ -21,25 +25,37 @@ fn main() {
 
 fn run() -> Result<()> {
     let args = env::args_os().skip(1).collect::<Vec<_>>();
-    let output = args[args.iter().position(|arg| arg == "-o").unwrap() + 1].clone();
     let config = Config::read_from_current_dir()?;
     let registry = Registry::new()?;
     let mut signals = register_signals()?;
+
+    let crate_root = crate_root()?;
+    let target = block_with_signals(&mut signals, true, run_drone_print_target)?;
+    let target = crate_root.join("target").join(target);
+    create_dir_all(&target)?;
+    let stage_one = target.join("layout.ld.1");
+    let stage_two = target.join("layout.ld.2");
     {
-        let script = registry.layout_ld(&config, false)?;
-        let linker = linker_command(script.as_ref(), &args, &[])?;
+        let stage_one_file = File::create(&stage_one)?;
+        let stage_two_file = File::create(&stage_two)?;
+        registry.layout_ld(&config, false, &stage_one_file)?;
+        registry.layout_ld(&config, true, &stage_two_file)?;
+    }
+
+    if let Some(output_position) = args.iter().position(|arg| arg == "-o") {
+        let linker = linker_command(stage_one.as_ref(), &args, &[])?;
+        block_with_signals(&mut signals, true, || run_command(linker))?;
+
+        let size = size_command(&args[output_position + 1])?;
+        let syms = block_with_signals(&mut signals, true, || run_size(size))?
+            .into_iter()
+            .map(|(name, size)| format!("--defsym=_{}_section_size={}", name, size))
+            .collect::<Vec<_>>();
+
+        let linker = linker_command(stage_two.as_ref(), &args, &syms)?;
         block_with_signals(&mut signals, true, || run_command(linker))?;
     }
-    let size = size_command(&output)?;
-    let syms = block_with_signals(&mut signals, true, || run_size(size))?
-        .into_iter()
-        .map(|(name, size)| format!("--defsym=_{}_section_size={}", name, size))
-        .collect::<Vec<_>>();
-    {
-        let script = registry.layout_ld(&config, true)?;
-        let linker = linker_command(script.as_ref(), &args, &syms)?;
-        block_with_signals(&mut signals, true, || run_command(linker))?;
-    }
+
     Ok(())
 }
 
@@ -69,4 +85,11 @@ fn run_size(mut command: Command) -> Result<HashMap<String, u32>> {
         }
     }
     Ok(map)
+}
+
+fn run_drone_print_target() -> Result<String> {
+    let mut command = Command::new("drone");
+    command.arg("print").arg("target");
+    let stdout = String::from_utf8(command.output()?.stdout)?;
+    Ok(stdout.trim().to_string())
 }